fix segfault in cursors tab of the debugger
When inspecting the driver or certain other widgets with the debugger,
the widget would sometimes have a driver of NULL, which could cause a
segfault. With this commit, the cursors tab is disabled in such cases.

This might be a band-aide fix? Maybe widgets are supposed to always be
guaranteed to have a driver?



git-svn-id: file:///home/dev/agar/trunk@10181 86a84d31-9702-0410-8f1e-a60cab7bd424