Fix horizontal mousewheel scrolling in `AG_Editable` and `AG_Console`.
In `AG_Textbox`, bind the horizontal scrollbar's increment to `fontMaxHeight`
and the vertical scrollbar's increment to `lineScrollAmount` such that the
scrolling behavior when the cursor is over the scrollbar areas matches the
scrolling behavior when it is over the `AG_Editable` area.

In `AG_Console`, reduce the clipping rectangle's width by 1px to avoid
visible clipping artifacts around the line boundary.



git-svn-id: file:///home/dev/agar/trunk@11020 86a84d31-9702-0410-8f1e-a60cab7bd424