Permit negative padding values to be emitted by the style compiler.
Negative padding may be used to condense widgets that support it
(such as `AG_Tlist`).

Handle the "padding" attribute on a per-item basis in `AG_Tlist`.

Raise the "padding-changed" event whenever the padding attribute
of an `AG_Widget` is modified. This allows widgets an opportunity
to invalidate any padding-dependent generated surface, as well as
to clamp padding values that cannot be handled by the widget.



git-svn-id: file:///home/dev/agar/trunk@11033 86a84d31-9702-0410-8f1e-a60cab7bd424